move env cli utils to own module
Move the get_filename and print_results function used only by the
conda env subcommand from conda/env/env.py to their own module
conda/env/env_util.py

This change removes the need to delay the imports done in print_results,
allowing the import to live at the top of the file.
