+++ conda/cli/main_env_create.py
@@ -103,7 +103,7 @@ def execute(args: Namespace, parser: ArgumentParser) -> int:
     from ..base.context import context, determine_target_prefix
     from ..core.prefix_data import PrefixData
     from ..env import specs
-    from ..env.env import get_filename, print_result
+    from ..env.env_util import get_filename, print_result
     from ..env.installers.base import get_installer
     from ..exceptions import InvalidInstaller
     from ..gateways.disk.delete import rm_rf
