+++ conda/cli/main_env_update.py
@@ -80,7 +80,7 @@ def execute(args: Namespace, parser: ArgumentParser) -> int:
     from ..base.context import context, determine_target_prefix
     from ..core.prefix_data import PrefixData
     from ..env import specs as install_specs
-    from ..env.env import get_filename, print_result
+    from ..env.env_util import get_filename, print_result
     from ..env.installers.base import get_installer
     from ..exceptions import CondaEnvException, InvalidInstaller
     from ..misc import touch_nonadmin
